requirements.txt
flask==3.1.0
opencv-contrib-python==4.10.0.84
numpy==1.26.4
Pillow==11.1.0
